Add endpoint to update event password and notify guests
A new endpoint has been added to the `EventoController` class to
allow updating the password of an event via an HTTP PUT request
at the route `"{id}/senha"`. The `UpdateEventPassword` method
checks for the new password, updates it using `_eventoService`,
and sends email notifications to all invited guests if applicable.

The `IEventoService` interface now includes the `UpdateEventPassword`
method, which is implemented in the `EventoService` class. This
method retrieves the event by its ID, updates its password hash,
and saves the changes to the repository.